==> 03_downstream_analysis/04_integration_with_annotation/02_fineTuning_scANVI.ipynb <==
import os
import sys

import json

import numpy as np
import scanpy as sc
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.metrics import classification_report

import scarches as sca
from scarches.dataset.trvae.data_handling import remove_sparsity
import matplotlib.pyplot as plt

from lightning.pytorch.callbacks import ModelCheckpoint

import torch

from dotenv import load_dotenv

from lightning.pytorch.loggers import WandbLogger
import wandb

import session_info
import warnings
from pyprojroot.here import here

#plt.style.use(['science','nature','no-latex'])
dpi_fig_save = 300
sc.set_figure_params(dpi=100, dpi_save=dpi_fig_save, vector_friendly=True)

# Setting some parameters
warnings.filterwarnings("ignore")

from sklearn.model_selection import StratifiedKFold

overwriteData = True
overwriteFigures = True

# Set random seed
random_seed = 42

import warnings
warnings.filterwarnings('ignore')

import scvi
scvi.settings.dl_num_workers = 0
scvi.settings.seed = random_seed

#torch.set_float32_matmul_precision('high')
#torch.set_printoptions(precision=3, sci_mode=False, edgeitems=7)
# torch.multiprocessing.set_sharing_strategy('file_system')

%load_ext autoreload
%autoreload 2

assert load_dotenv()

workingDir = here('03_downstream_analysis/04_integration_with_annotation/')
workingDir

class CustomWandbLogger(WandbLogger):
    @property
    def save_dir(self):
        """Gets the save directory.

        Returns:
            The path to the save directory.

        """
        return self.experiment.dir

scvi.__version__

# Loading data


# Load the h5ad file
adata = sc.read_h5ad(here("03_downstream_analysis/04_integration_with_annotation/04_MAIN_geneUniverse_noRBCnPlatelets.h5ad"))#, 
                     #backed='r+', chunk_size=50000)


adata.obs['binned_age'] = adata.obs['binned_age'].astype(str)

adata.obs['binned_age_cod'] = adata.obs.binned_age.cat.codes
adata.obs['chemistry_cod'] = adata.obs.chemistry.cat.codes
adata.obs['libraryID_cod'] = adata.obs.libraryID.cat.codes
adata.obs['studyID_cod'] = adata.obs.studyID.cat.codes
adata.obs['sex_cod'] = adata.obs.sex.cat.codes
adata.obs['Level1_cod'] = adata.obs.Level1.cat.codes.astype(str)

#### Preparing scANVI training

scvi_model = sca.models.SCVI.load(here(f"{workingDir}/results/scVI_model_pretreined_noRBCnPlat/"), adata=adata) 

### Fine tuning with scANVI

**Parameters**

scANVI_trainer_kwargs = dict(
    n_samples_per_label = None,
    check_val_every_n_epoch = None,
    train_size = 0.8,
    validation_size = 0.2,
    shuffle_set_split = True,
    checkpointing_monitor = 'elbo_validation',
    early_stopping_monitor = 'reconstruction_loss_validation',
    early_stopping_patience = 2,
    early_stopping_min_delta=0.1,
    early_stopping = True,
    max_epochs = 1000,
)
# https://docs.scvi-tools.org/en/stable/api/reference/scvi.train.TrainingPlan.html#scvi.train.TrainingPlan
plan_kwargs = dict(
    lr = 5e-5,
    #reduce_lr_on_plateau = True
)
datasplitter_kwargs = dict(pin_memory=False)
scanvi_parameter_dict = scANVI_trainer_kwargs | plan_kwargs | datasplitter_kwargs

run_name = f"MAINobj_scANVI_fineTuning_lowLR_noRBCnPlat"
run_name

scanvi_model = sca.models.SCANVI.from_scvi_model(scvi_model, unlabeled_category = "unknown")
scanvi_model

logger = CustomWandbLogger(name = run_name, project='inflammation_atlas_R1_scANVI', config = scanvi_parameter_dict)

model_checkpoint = ModelCheckpoint(monitor='elbo_validation', every_n_epochs=1, save_last=True, save_top_k = -1, 
                                   dirpath=f"{workingDir}/results/scANVI_model_fineTuned_lowLR_noRBCnPlat_checkpoints/")

try:
    scanvi_model.train(logger=logger, plan_kwargs = plan_kwargs, datasplitter_kwargs=datasplitter_kwargs, 
                       enable_checkpointing=True, callbacks=[model_checkpoint], 
                       **scANVI_trainer_kwargs) #SAME parameter as scVI
    scanvi_model.save(here(f"{workingDir}/results/scANVI_model_fineTuned_lowLR_noRBCnPlat/"),
               overwrite = True, 
               save_anndata = False)
    scanvi_emb = scanvi_model.get_latent_representation(adata=adata)
    np.savez_compressed(file = str(here(f"{workingDir}/results/scANVI_model_fineTuned_lowLR_noRBCnPlat/scANVI_embedding.npz")), arr=scanvi_emb)
    
except Exception as e:
    print(f"An error occurred: {e}")
    scanvi_model.save(here(f"{workingDir}/results/scANVI_model_fineTuned_lowLR_noRBCnPlat_WITHERRORS/"), 
                      overwrite=True, 
                      save_anndata=False)

wandb.finish()

if overwriteData:
    scanvi_model.save(here(f"{workingDir}/results/scANVI_model_pretreined_noRBCnPlat/"),
               overwrite = True, 
               save_anndata = False)

scanvi_emb = scanvi_model.get_latent_representation(adata=adata)

np.savez_compressed(file = str(here(f"{workingDir}/results/scANVI_model_pretreined_noRBCnPlat/scANVI_embedding.npz")), arr=scanvi_emb)
