==> 03_downstream_analysis/04_integration_with_annotation/03_compare_latent_embeddings.ipynb <==
### Intro

The aim of the this notebook (and the 04_benchmark_integration) is to test the integration comparing non-integrated PC, scANVI embedding and PC obtained from scANVI normalized expression matrix.

Here, we will generate an adata object including three embedded space as explained above.

For each embedding we will use the same number of dimension (30)

import scanpy as sc

import pandas as pd
import numpy as np

import scvi

from pyprojroot import here
from glob import glob
import os

### Loading raw non-integrated object

This object was integrated using scANVI

adata = sc.read_h5ad(here("03_downstream_analysis/02_gene_universe_definition/results/04_MAIN_geneUniverse_noRBCnPlatelets.h5ad"))

adata.obs['binned_age'] = adata.obs['binned_age'].astype(str)

### Loading scANVI model

scanvi_model = scvi.model.SCANVI.load(here('03_downstream_analysis/04_integration_with_annotation/results/scANVI_model_fineTuned_lowLR_noRBCnPlat'), 
                                      adata=adata) 

## Generating embeddings

### scANVI latent dimension

adata.obsm['X_scANVI_latent'] = scanvi_model.get_latent_representation(adata=adata)

### PCA from RAW count
#### Normalization

sc.pp.normalize_total(adata, target_sum=1e4)
sc.pp.log1p(adata)

#### Computing PCA

sc.tl.pca(adata, n_comps = 30, )

adata.obsm['X_pca_unintegrated'] = adata.obsm['X_pca'].copy()

del adata.obsm['X_pca']

### PCA from scANVI normalized expression

For the third embedding we have to load the  normalized count data obtained using get_normalized_expression() from scANVI model.

We generated the those matrices for each cell-type to make loading easier

**Generating one adata by merging each cell-types**

ctDATApath_list = glob(str(here('03_downstream_analysis/04_integration_with_annotation/results/normalized_adatas_nextflow/cellType_adata_merged/*.log1p.h5ad')))

ctADATA_list = []
for p in ctDATApath_list:
    print(os.path.basename(p))
    ctADATA_list.append(sc.read_h5ad(p))

adataCTmerged = sc.concat(ctADATA_list)[adata.obs_names,:]

assert all(adataCTmerged.obs_names == adata.obs_names)

#### Computing PCA

sc.tl.pca(adataCTmerged, n_comps = 30)

adata.obsm['X_pca_scANVI_normalized_exp'] = adataCTmerged.obsm['X_pca'].copy()

### scGen latent dimension

**Load adata with corrected latent and PCs**

scGen_emb_adata = sc.read_h5ad(here('03_downstream_analysis/04_integration_with_annotation/results/02_scGen_corr_latent_expression_noRBCnPlat.h5ad'))
scGen_emb_adata

scGen_emb_adata

# Check if index of the two adata perfectly match
assert(all(scGen_emb_adata.obs_names == adata.obs_names))

adata.obsm['scgen_corrected_latent'] = scGen_emb_adata.layers['scgen_corrected_latent']
adata.obsm['scgen_corrected_expression_pc'] = scGen_emb_adata.layers['scgen_corrected_expression_pc']

### Harmony latent dimension

**Load adata with corrected PCs**

harmony_adata = sc.read_h5ad(here('03_downstream_analysis/04_integration_with_annotation/results/02_Harmony_correctedPCs_noRBCnPlat.h5ad'))
harmony_adata

# Check if index of the two adata perfectly match
assert(all(harmony_adata.obs_names == adata.obs_names))

adata.obsm['X_pca_harmony'] = harmony_adata.X
adata

## Saving the obtained object

adata.write(here('03_downstream_analysis/04_integration_with_annotation/results/03_MAIN_geneUniverse_noRBCnPlatelets_embeddings.h5ad'), compression='gzip')

adata = sc.read_h5ad(here('03_downstream_analysis/04_integration_with_annotation/results/03_MAIN_geneUniverse_noRBCnPlatelets_embeddings.h5ad'))
